+++ client/src/settings.js
@@ -3,11 +3,12 @@
 module.exports = {
   clientSideRenderingFail: 'Client-side rendering failed',
   configurationNotValid: 'Configuration is not valid: ',
-  connectionError: '{0} when connecting to {1}',
+  connectionError: 'request {0} failed ({1})',
   emptyResponse: 'Empty response',
   genericError: 'Internal client error',
   legacyComponent: 'The component can\'t be rendered because it was published with an older OC version',
   registriesEmpty: 'registries must contain at least one endpoint',
   registriesIsNotObject: 'registries must be an object',
-  serverSideRenderingFail: 'Server-side rendering failed: {0}'
+  serverSideRenderingFail: 'Server-side rendering failed: {0}',
+  warmupFailed: 'Error warming up oc-client: request {0} failed ({1})'
 };