+++ src/registry/domain/options-sanitiser.js
@@ -34,5 +34,15 @@ module.exports = function(input){
     options.dependencies = dependenciesResolver(options);
   }
 
+  if(!_.isBoolean(options.hotReloading)){
+    options.hotReloading = !!options.local;
+  }
+
+  if(_.isUndefined(options.verbosity)){
+    options.verbosity = 0;
+  }
+
+  options.port = process.env.PORT || options.port;
+
   return options;
 };