+++ test/acceptance/client.js
@@ -3,6 +3,7 @@
 var cheerio = require('cheerio');
 var expect = require('chai').expect;
 var path = require('path');
+var _ = require('underscore');
 
 describe('The node.js OC client', function(){
 
@@ -34,6 +35,12 @@ describe('The node.js OC client', function(){
     };
   };
 
+  var getRegExpFromJson = function(x){
+    return JSON.stringify(x)
+      .replace(/\+/g, '\\+')
+      .replace(/\[/g, '\\[');
+  };
+
   describe('when initialised providing registries properties', function(){
 
     before(function(done){
@@ -46,6 +53,111 @@ describe('The node.js OC client', function(){
     after(function(done){ registry.close(done); });
 
     describe('when rendering 2 components', function(){
+      describe('when components require params', function(){
+        describe('when each component requires different params', function(){
+          var $components;
+          var $errs;
+          before(function(done){
+            client.renderComponents([{
+              name: 'welcome',
+              parameters: {
+                firstName: 'Jane',
+                lastName: 'Marple'
+              }
+            }, {
+              name: 'welcome',
+              parameters: {
+                firstName: 'Hercule',
+                lastName: 'Poirot'
+              }
+            }], { container: false, renderInfo: false }, function(err, html){
+              $errs = err;
+              $components = html;
+              done();
+            });
+          });
+
+          it('should return rendered contents', function(){
+            expect($components[0]).to.contain('hi Jane Marple');
+            expect($components[1]).to.contain('hi Hercule Poirot');
+          });
+
+          it('should return null errors', function () {
+            expect($errs).to.be.null;
+          });
+        });
+
+        describe('when each component requires the same params', function(){
+          var $components;
+          var $errs;
+          before(function(done){
+            client.renderComponents([{
+              name: 'welcome'
+            }, {
+              name: 'welcome'
+            }], {
+              container: false,
+              parameters: {
+                firstName: 'Jane',
+                lastName: 'Marple'
+              },
+              renderInfo: false
+            }, function(err, html){
+              $errs = err;
+              $components = html;
+              done();
+            });
+          });
+
+          it('should return rendered contents', function(){
+            expect($components[0]).to.contain('hi Jane Marple');
+            expect($components[1]).to.contain('hi Jane Marple');
+          });
+
+          it('should return null errors', function () {
+            expect($errs).to.be.null;
+          });
+        });
+
+        describe('when components have some common parameters and some different', function(){
+          var $components;
+          var $errs;
+          before(function(done){
+            client.renderComponents([{
+              name: 'welcome',
+              parameters: {
+                lastName: 'Poirot'
+              }
+            }, {
+              name: 'welcome',
+              parameters: {
+                firstName: 'Jane'
+              }
+            }], {
+              container: false,
+              parameters: {
+                firstName: 'Hercule',
+                lastName: 'Marple'
+              },
+              renderInfo: false
+            }, function(err, html){
+              $errs = err;
+              $components = html;
+              done();
+            });
+          });
+
+          it('should return rendered contents', function(){
+            expect($components[0]).to.contain('hi Hercule Poirot');
+            expect($components[1]).to.contain('hi Jane Marple');
+          });
+
+          it('should return null errors', function () {
+            expect($errs).to.be.null;
+          });
+        });
+      });
+
       describe('when rendering both on the server-side', function(){
         var $components;
         var $errs;
@@ -151,22 +263,32 @@ describe('The node.js OC client', function(){
         });
       });
 
-      describe('when there\'s error in one of them', function(){
+      describe('when there are errors in some of them', function(){
         var $errs;
         before(function(done){
           client.renderComponents([{
             name: 'hello-world-i-dont-exist'
           }, {
             name: 'no-containers'
-          }], { container: false, renderInfo: false }, function(err, html){
+          }, {
+            name: 'errors-component',
+            parameters: {
+              errorType: '500'
+            }
+          }], {
+            container: false,
+            renderInfo: false,
+            disableFailoverRendering: true
+          }, function(err, html){
             $errs = err;
             done();
           });
         });
 
-        it('should return an error for the component with error', function(){
+        it('should return an error for each component with error', function(){
           expect($errs[0].toString()).to.be.equal('Error: Server-side rendering failed: Component "hello-world-i-dont-exist" not found on local repository (404)');
           expect($errs[1]).to.be.null;
+          expect($errs[2].toString()).to.be.equal('Error: Server-side rendering failed: Component execution error: An error happened (500)');
         });
       });
     });
@@ -216,6 +338,24 @@ describe('The node.js OC client', function(){
 
     describe('when server-side rendering an existing component linked to a non responsive registry', function(){
 
+      var expectedRequest = {
+        url: 'http://localhost:1234',
+        method: 'post',
+        headers: { 
+          accept: 'application/vnd.oc.unrendered+json',
+          'user-agent': 'oc-client-(.*?)'
+        },
+        timeout: 5,
+        json: true,
+        body: {
+          components: [{
+            name: 'hello-world',
+            version: '~1.0.0'
+          }],
+          parameters: {}
+        }
+      };
+
       describe('when client-side failover rendering disabled', function(){
 
         var error;
@@ -235,7 +375,11 @@ describe('The node.js OC client', function(){
         });
 
         it('should contain the error details', function(){
-          expect(error.toString()).to.match(/Error\: Server-side rendering failed: Error: (.*?) when connecting to http\:\/\/localhost\:1234 \(500\)/g);
+
+          var exp = getRegExpFromJson(expectedRequest),
+              expected = new RegExp('Error: Server-side rendering failed: request ' + exp + ' failed \\(Error: connect ECONNREFUSED(.*?)\\)');
+
+          expect(error.toString()).to.match(expected);
         });
       });
 
@@ -268,7 +412,11 @@ describe('The node.js OC client', function(){
         });
 
         it('should contain the error details', function(){
-          expect(error.toString()).to.match(/Error\: Server-side rendering failed: Error: (.*?) when connecting to http\:\/\/localhost\:1234 \(500\)/g);
+
+          var exp = getRegExpFromJson(expectedRequest),
+              expected = new RegExp('Error: Server-side rendering failed: request ' + exp + ' failed \\(Error: connect ECONNREFUSED(.*?)\\)');
+
+          expect(error.toString()).to.match(expected);
         });
       });
 
@@ -307,7 +455,35 @@ describe('The node.js OC client', function(){
         });
 
         it('should contain the error details', function(){
-          expect(error.toString()).to.match(/Error\: Server-side rendering failed: Error: (.*?) when connecting to http\:\/\/localhost\:1234 \(500\)/g);
+
+          var expectedRequestWithExtraParams = {
+            url: 'http://localhost:1234',
+            method: 'post',
+            headers: {
+              'accept-language': 'da, en-gb;q=0.8, en;q=0.7',
+              accept: 'application/vnd.oc.unrendered+json',
+              'user-agent': 'oc-client-(.*?)'
+            },
+            timeout: 5,
+            json: true,
+            body: {
+              components: [{
+                name: 'hello-world',
+                version: '~1.0.0',
+                parameters: {
+                  hi: 'john'
+                }
+              }],
+              parameters: {
+                hi: 'john'
+              }
+            }
+          };
+
+          var exp = getRegExpFromJson(expectedRequestWithExtraParams),
+              expected = new RegExp('Error: Server-side rendering failed: request ' + exp + ' failed \\(Error: connect ECONNREFUSED(.*?)\\)');
+
+          expect(error.toString()).to.match(expected);
         });
       });
 
@@ -341,12 +517,68 @@ describe('The node.js OC client', function(){
         });
 
         it('should contain the error details', function(){
-          expect(error.toString()).to.match(/Error\: Server-side rendering failed: Error: (.*?) when connecting to http\:\/\/localhost\:1234 \(500\)/g);
+          var exp = getRegExpFromJson(expectedRequest),
+              expected = new RegExp('Error: Server-side rendering failed: request ' + exp + ' failed \\(Error: connect ECONNREFUSED(.*?)\\)');
+
+          expect(error.toString()).to.match(expected);
         });
       });
     });
 
     describe('when server-side rendering an existing component linked to a responsive registry', function(){
+
+      describe('when the component times-out', function(){
+
+        var error, result;
+
+        var expectedRequest = {
+          url: 'http://localhost:3030',
+          method: 'post',
+          headers: { 
+            accept: 'application/vnd.oc.unrendered+json',
+            'user-agent': 'oc-client-(.*?)'
+          },
+          timeout: 0.01,
+          json: true,
+          body: {
+            components: [{
+              name: 'errors-component',
+              parameters: {
+                errorType: 'timeout',
+                timeout: 1000
+              }
+            }],
+            parameters: {
+              errorType: 'timeout',
+              timeout: 1000
+            }
+          }
+        };
+
+        before(function(done){
+          client.renderComponent('errors-component', {
+            parameters: { errorType: 'timeout', timeout: 1000 },
+            timeout: 0.01,
+            disableFailoverRendering: true
+          }, function(err, html){
+            error = err;
+            result = html;
+            done();
+          });
+        });
+
+        it('should contain a blank html response', function(){
+          expect(result).to.eql('');
+        });
+
+        it('should contain the error details', function(){
+
+          var exp = getRegExpFromJson(expectedRequest),
+              expected = new RegExp('Error: Server-side rendering failed: request ' + exp + ' failed \\(timeout\\)');
+
+          expect(error.toString()).to.match(expected);
+        });
+      });
 
       describe('when container option = true', function(){
         var error;