+++ test/fixtures/components/welcome-with-optional-parameters/_package/server.js
@@ -1,0 +1,1 @@
+"use strict";module.exports.data=function(a,s){s(null,{firstName:a.params.firstName,lastName:a.params.lastName,nick:a.params.nick})};