+++ test/unit/registry-domain-repository.js
@@ -269,11 +269,13 @@ describe('registry : domain : repository', function(){
         expect(response.result).to.eql([
           'container-with-multiple-nested',
           'container-with-nested',
+          'errors-component',
           'handlebars3-component',
           'hello-world',
           'language',
           'no-containers',
           'welcome',
+          'welcome-with-optional-parameters',
           'oc-client'
         ]);
       });